Integrate highlight-blockquote.js into the application.
