+++ src/badge-highlight/src/index.ts
@@ -1,1 +1,1 @@
-export * from './hightlight';
+
