+++ src/mainwindow.cpp
@@ -629,6 +629,7 @@ MainWindow::htmlContent() const
 		"<meta charset=\"utf-8\">\n"
 		"<head>\n"
 		"  <script src=\"qrc:/qtwebchannel/qwebchannel.js\"></script>\n"
+		"  <script src=\"qrc:/res/hightlight-blockquote.js\"></script>\n"
 		"  <link rel=\"stylesheet\" href=\"qrc:/res/css/github.css\">\n"
 		"  <link rel=\"stylesheet\" href=\"qrc:/res/css/github-markdown.css\">\n"
 		"  <script src=\"qrc:/res/highlight.js\"></script>\n"
@@ -656,6 +657,7 @@ MainWindow::htmlContent() const
 		"         strict : false,\n"
 		"         trust : true\n"
 		"     });\n"
+		"     replaceBadges(document.body);\n"
 		"  }\n"
 		"\n"
 		"  new QWebChannel(qt.webChannelTransport,\n"
